Script to dump public task information
Dumps to a json file for reporting and analytics that several
departments have requested.  All security information is
derived from the security on the task, though project view policy
is also respected.

This is very specific as several columns in tables contain sensitive
or potentially sensitive information even on public tasks such as
mailKeys or hidden history.

Change-Id: I9e20d77a5642757c96f32f40914528b08104abbd