use dbslave for public dump
This changes things up for the public dump job
to more sanely work with mysql connections.  Ideally this
is all redone and we ditch all of the migration logic
that was never meant to live this long.  In practice at
least this does more sane things even though it all needs to be
scrapped and passing db con objects around in this way is odd.

Change-Id: Ib26179676c08996ea158c15f6d397b89987eeb2e